@@ -1,22 +1,28 @@
 {
   "name": "linkify",
-  "version": "3.0.5",
+  "version": "4.0.0-beta.6",
   "description": "Intelligent link recognition, made easy",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/Hypercontext/linkifyjs.git"
   },
   "scripts": {
-    "clean": "rm -rf dist",
-    "coverage": "nyc report --reporter=text-lcov | coveralls",
     "build": "npm run build --workspaces",
     "build:ci": "for dir in `ls packages` ; do cd packages/${dir} && npm run build && cd ../.. ; done",
-    "copy": "copyfiles -f packages/*/dist/*.js LICENSE dist",
-    "dist": "run-s clean build copy",
-    "dist:ci": "run-s clean build:ci copy",
-    "lint": "eslint *.js packages/**/*.js test/**/*.js",
-    "test": "nyc --reporter=lcov --reporter=text mocha test/index.js",
-    "test:ci": "karma start test/ci.conf.js --single-run"
+    "clean": "rm -rf dist",
+    "copy": "copyfiles -u 3 packages/*/dist/*.js packages/*/dist/*/LICENSE dist",
+    "copy:license": "copyfiles LICENSE dist",
+    "coverage": "nyc report --reporter=text-lcov | coveralls",
+    "dist": "run-s clean prepack copy copy:license",
+    "dist:postpublish": "run-s clean copy copy:license",
+    "lint": "eslint *.js packages/**/*.js test/**/*.js tasks/**/*.js",
+    "prepack": "npm run prepack --workspaces",
+    "test": "mocha test/index.js",
+    "test:coverage": "nyc --reporter=lcov --reporter=text mocha test/index.js",
+    "test:ci": "run-s test:ci1 test:ci2",
+    "test:ci1": "karma start test/ci1.conf.js --single-run",
+    "test:ci2": "karma start test/ci2.conf.js --single-run",
+    "tlds": "node tasks/update-tlds.js"
   },
   "author": "Hypercontext",
   "license": "MIT",
@@ -26,30 +32,31 @@
     "@babel/eslint-plugin": "^7.13.10",
     "@babel/preset-env": "^7.13.10",
     "@babel/register": "^7.13.8",
-    "@nfrasser/simple-html-tokenizer": "==0.5.11-1",
+    "@nfrasser/simple-html-tokenizer": "==0.5.11-2",
     "@rollup/plugin-babel": "^5.3.0",
-    "@rollup/plugin-commonjs": "^17.1.0",
-    "@rollup/plugin-node-resolve": "^11.2.0",
+    "@rollup/plugin-commonjs": "^21.0.1",
+    "@rollup/plugin-node-resolve": "^13.0.6",
     "chai": "^4.3.4",
     "copyfiles": "^2.4.1",
     "coveralls": "^3.1.0",
-    "eslint": "^7.22.0",
-    "eslint-plugin-mocha": "^8.1.0",
-    "glob": "^7.1.6",
+    "eslint": "^8.3.0",
+    "eslint-plugin-mocha": "^10.0.3",
+    "glob": "^8.0.1",
     "jquery": "^3.6.0",
     "jsdom": "^11.12.0",
-    "karma": "6.3.16",
+    "karma": "^6.3.16",
     "karma-browserstack-launcher": "^1.6.0",
     "karma-chrome-launcher": "^3.1.0",
     "karma-firefox-launcher": "^2.1.0",
     "karma-qunit": "^4.1.2",
-    "mocha": "^8.3.2",
+    "mocha": "^9.2.2",
     "mocha-lcov-reporter": "^1.3.0",
     "npm-run-all": "^4.1.5",
     "nyc": "^15.1.0",
+    "punycode": "^2.1.1",
     "qunit": "^2.14.1",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
     "request": "^2.88.2",
     "rollup": "^2.42.1",
     "rollup-plugin-terser": "^7.0.2",
@@ -60,6 +67,8 @@
     "node": ">=8"
   },
   "workspaces": [
+    "./packages/linkifyjs",
+    "./packages/linkify-plugin-*",
     "./packages/*"
   ]
 }
