@@ -1,7 +1,76 @@
 /* eslint-disable mocha/no-setup-in-describe */
-const linkify = require('linkifyjs/src/linkify');
+import { expect } from 'chai';
+import * as linkify from 'linkifyjs/src/linkify';
+
+const TicketToken = linkify.createTokenClass('ticket', { isLink: true });
+
+/**
+ * @type import('linkifyjs').Plugin
+ */
+const ticketPlugin = ({ scanner, parser }) => {
+	const { POUND, groups } = scanner.tokens;
+	const Hash = parser.start.tt(POUND);
+	const Ticket = new linkify.State(TicketToken);
+	Hash.ta(groups.numeric, Ticket);
+};
 
 describe('linkifyjs', () => {
+	describe('registerPlugin', () => {
+		beforeEach(() => {
+			linkify.registerPlugin('ticket', ticketPlugin);
+		});
+
+		it('Detects tickets after applying', () => {
+			expect(linkify.test('#123', 'ticket')).to.be.ok;
+		});
+
+		it('Logs a warning if registering same plugin twice', () => {
+			linkify.registerPlugin('ticket', ticketPlugin);
+			expect(linkify.test('#123', 'ticket')).to.be.ok;
+		});
+
+		it('Logs a warning if already initialized', () => {
+			linkify.init();
+			linkify.registerPlugin('ticket2', ticketPlugin);
+		});
+	});
+
+	describe('registerCustomProtocol', () => {
+		beforeEach(() => {
+			linkify.registerCustomProtocol('instagram', true);
+			linkify.registerCustomProtocol('view-source');
+		});
+
+		it('Detects basic protocol', () => {
+			expect(linkify.test('instagram:user/nfrasser', 'url')).to.be.ok;
+		});
+
+		it('Detects basic protocol with slash slash', () => {
+			expect(linkify.test('instagram://user/nfrasser', 'url')).to.be.ok;
+		});
+
+		it('Detects compound protocol', () => {
+			expect(linkify.test('view-source://http://github.com/', 'url')).to.be.ok;
+		});
+
+		it('Does not detect protocol with non-optional //', () => {
+			expect(linkify.test('view-source:http://github.com/', 'url')).to.not.be.ok;
+		});
+
+		it('Does not detect custom protocol if already initialized', () => {
+			linkify.init();
+			linkify.registerCustomProtocol('fb');
+			expect(linkify.test('fb://feed')).to.not.be.ok;
+		});
+
+		it('Throws error when protocol has invalid format', () => {
+			expect(() => linkify.registerCustomProtocol('-')).to.throw();
+			expect(() => linkify.registerCustomProtocol('-fb')).to.throw();
+			expect(() => linkify.registerCustomProtocol('fb-')).to.throw();
+			expect(() => linkify.registerCustomProtocol('git+https')).to.throw(); // this may work in the future
+		});
+	});
+
 	describe('tokenize', () => {
 		it('is a function', () => {
 			expect(linkify.tokenize).to.be.a('function');
@@ -15,8 +84,65 @@ describe('linkifyjs', () => {
 		it('is a function', () => {
 			expect(linkify.find).to.be.a('function');
 		});
-		it('takes a single argument', () => {
-			expect(linkify.find.length).to.be.eql(1); // type is optional
+
+		it('Find nothing in an empty string', () => {
+			expect(linkify.find('')).to.deep.eql([]);
+		});
+
+		it('Find nothing in a string with no links', () => {
+			expect(linkify.find('Hello World!')).to.deep.eql([]);
+		});
+
+		it('Find the link', () => {
+			expect(linkify.find('hello.world!')).to.deep.eql([{
+				type: 'url',
+				value: 'hello.world',
+				href: 'http://hello.world',
+				isLink: true,
+				start: 0,
+				end: 11
+			}]);
+		});
+
+		it('Find the link of the specific type', () => {
+			expect(linkify.find('For help with github.com, please contact support@example.com', 'email')).to.deep.eql([{
+				type: 'email',
+				value: 'support@example.com',
+				href: 'mailto:support@example.com',
+				isLink: true,
+				start: 41,
+				end: 60
+			}]);
+		});
+
+		it('Finds with opts', () => {
+			expect(linkify.find('Does www.truncate.com work with truncate?', { truncate: 10 })).to.deep.eql([
+				{
+					type: 'url',
+					value: 'www.trunca…',
+					isLink: true,
+					href: 'http://www.truncate.com',
+					start: 5,
+					end: 21
+				}
+			]);
+		});
+
+		it('Finds type and opts', () => {
+			expect(linkify.find('Does www.truncate.com work with example@truncate.com?', 'email', { truncate: 10 })).to.deep.eql([
+				{
+					type: 'email',
+					value: 'example@tr…',
+					isLink: true,
+					href: 'mailto:example@truncate.com',
+					start: 32,
+					end: 52
+				}
+			]);
+		});
+
+		it('Throws on ambiguous invocation', () => {
+			expect(() => linkify.find('Hello.com', { type: 'email' }, { truncate: 10 })).to.throw();
 		});
 	});
 
@@ -38,7 +164,7 @@ describe('linkifyjs', () => {
 			['test+4@uwaterloo.ca', true],
 			['test+4@uwaterloo.ca', false, 'url'],
 			['test+4@uwaterloo.ca', true, 'email'],
-			['mailto:test+5@uwaterloo.ca', true, 'email'],
+			['mailto:test+5@uwaterloo.ca', true, 'url'],
 			['t.co', true],
 			['t.co g.co', false], // can only be one
 			['test@g.co t.co', false] // can only be one
@@ -46,9 +172,6 @@ describe('linkifyjs', () => {
 
 		it('is a function', () => {
 			expect(linkify.test).to.be.a('function');
-		});
-		it('takes a single argument', () => {
-			expect(linkify.test.length).to.be.eql(1); // type is optional
 		});
 
 		let testName;
