@@ -1,13 +1,10 @@
-/**
-	Linkify a HTML DOM node
-*/
 import { tokenize, Options } from 'linkifyjs';
 
 const HTML_NODE = 1, TXT_NODE = 3;
 
 /**
  * @param {HTMLElement} parent
- * @param {Text | HTMLElement} oldChild
+ * @param {Text | HTMLElement | ChildNode} oldChild
  * @param {Array<Text | HTMLElement>} newChildren
  */
 function replaceChildWithChildren(parent, oldChild, newChildren) {
@@ -20,61 +17,22 @@ function replaceChildWithChildren(parent, oldChild, newChildren) {
 }
 
 /**
- * @param {MultiToken[]} tokens
- * @param {Object} opts
- * @param {Document} doc A
+ * @param {import('linkifyjs').MultiToken[]} tokens
+ * @param {import('linkifyjs').Options} options
+ * @param {Document} doc Document implementation
  * @returns {Array<Text | HTMLElement>}
  */
-function tokensToNodes(tokens, opts, doc) {
+function tokensToNodes(tokens, options, doc) {
 	const result = [];
 	for (let i = 0; i < tokens.length; i++) {
 		const token = tokens[i];
-		if (token.t === 'nl' && opts.nl2br) {
+		if (token.t === 'nl' && options.get('nl2br')) {
 			result.push(doc.createElement('br'));
-			continue;
-		} else if (!token.isLink || !opts.check(token)) {
+		} else if (!token.isLink || !options.check(token)) {
 			result.push(doc.createTextNode(token.toString()));
-			continue;
+		} else {
+			result.push(options.render(token));
 		}
-
-		const {
-			formatted,
-			formattedHref,
-			tagName,
-			className,
-			target,
-			rel,
-			events,
-			attributes,
-		} = opts.resolve(token);
-
-		// Build the link
-		const link = doc.createElement(tagName);
-		link.setAttribute('href', formattedHref);
-
-		if (className) { link.setAttribute('class', className); }
-		if (target) { link.setAttribute('target', target); }
-		if (rel) { link.setAttribute('rel', rel); }
-
-		// Build up additional attributes
-		if (attributes) {
-			for (const attr in attributes) {
-				link.setAttribute(attr, attributes[attr]);
-			}
-		}
-
-		if (events) {
-			for (const event in events) {
-				if (link.addEventListener) {
-					link.addEventListener(event, events[event]);
-				} else if (link.attachEvent)  {
-					link.attachEvent('on' + event, events[event]);
-				}
-			}
-		}
-
-		link.appendChild(doc.createTextNode(formatted));
-		result.push(link);
 	}
 
 	return result;
@@ -82,22 +40,20 @@ function tokensToNodes(tokens, opts, doc) {
 
 /**
  * Requires document.createElement
- * @param {HTMLElement} element
- * @param {Object} opts
+ * @param {HTMLElement | ChildNode} element
+ * @param {import('linkifyjs').Options} options
  * @param {Document} doc
  * @returns {HTMLElement}
  */
-function linkifyElementHelper(element, opts, doc) {
+function linkifyElementHelper(element, options, doc) {
 
 	// Can the element be linkified?
 	if (!element || element.nodeType !== HTML_NODE) {
 		throw new Error(`Cannot linkify ${element} - Invalid DOM Node type`);
 	}
 
-	const { ignoreTags } = opts;
-
 	// Is this element already a link?
-	if (element.tagName === 'A' || ignoreTags.indexOf(element.tagName) >= 0) {
+	if (element.tagName === 'A' || options.ignoreTags.indexOf(element.tagName) >= 0) {
 		// No need to linkify
 		return element;
 	}
@@ -109,7 +65,7 @@ function linkifyElementHelper(element, opts, doc) {
 
 		switch (childElement.nodeType) {
 		case HTML_NODE:
-			linkifyElementHelper(childElement, opts, doc);
+			linkifyElementHelper(childElement, options, doc);
 			break;
 		case TXT_NODE: {
 			str = childElement.nodeValue;
@@ -120,7 +76,7 @@ function linkifyElementHelper(element, opts, doc) {
 				break;
 			}
 
-			nodes = tokensToNodes(tokens, opts, doc);
+			nodes = tokensToNodes(tokens, options, doc);
 
 			// Swap out the current child for the set of nodes
 			replaceChildWithChildren(element, childElement, nodes);
@@ -139,15 +95,36 @@ function linkifyElementHelper(element, opts, doc) {
 }
 
 /**
+ * @param {Document} doc The document implementaiton
+ */
+function getDefaultRender(doc) {
+	return ({ tagName, attributes, content, eventListeners }) => {
+		const link = doc.createElement(tagName);
+		for (const attr in attributes) {
+			link.setAttribute(attr, attributes[attr]);
+		}
+
+		if (eventListeners && link.addEventListener) {
+			for (const event in eventListeners) {
+				link.addEventListener(event, eventListeners[event]);
+			}
+		}
+
+		link.appendChild(doc.createTextNode(content));
+		return link;
+	};
+}
+
+/**
  * Recursively traverse the given DOM node, find all links in the text and
  * convert them to anchor tags.
  *
  * @param {HTMLElement} element A DOM node to linkify
- * @param {Object} opts linkify options
+ * @param {import('linkifyjs').Opts} [opts] linkify options
  * @param {Document} [doc] (optional) window.document implementation, if differs from global
  * @returns {HTMLElement}
  */
-export default function linkifyElement(element, opts, doc = null) {
+export default function linkifyElement(element, opts = null, doc = null) {
 	try {
 		doc = doc || document || window && window.document || global && global.document;
 	} catch (e) { /* do nothing for now */ }
@@ -160,10 +137,18 @@ export default function linkifyElement(element, opts, doc = null) {
 		);
 	}
 
-	opts = new Options(opts);
-	return linkifyElementHelper(element, opts, doc);
+	const options = new Options(opts, getDefaultRender(doc));
+	return linkifyElementHelper(element, options, doc);
 }
 
-// Maintain reference to the recursive helper to cache option-normalization
+// Maintain reference to the recursive helper and option-normalization for use
+// in linkify-jquery
 linkifyElement.helper = linkifyElementHelper;
-linkifyElement.normalize = (opts) => new Options(opts);
+linkifyElement.getDefaultRender = getDefaultRender;
+
+/**
+ * @param {import('linkifyjs').Opts | import('linkifyjs').Options} opts
+ * @param {Document} doc
+ */
+linkifyElement.normalize = (opts, doc) => new Options(opts, getDefaultRender(doc));
+
